generalize reservoir output profile syncing in preparation to potentially apply to other examples
