+++ tests/examples/example5.out
@@ -6,8 +6,8 @@ Simulation Metadata
 ----------------------
  GEOPHIRES Version: 3.11.4
  Simulation Date: 2026-01-24
- Simulation Time:  14:41
- Calculation Time:      0.078 sec
+ Simulation Time:  15:04
+ Calculation Time:      0.077 sec
 
                            ***SUMMARY OF RESULTS***
