+++ tests/examples/example5b.out
@@ -6,7 +6,7 @@ Simulation Metadata
 ----------------------
  GEOPHIRES Version: 3.11.4
  Simulation Date: 2026-01-24
- Simulation Time:  14:41
+ Simulation Time:  15:04
  Calculation Time:      0.082 sec
 
                            ***SUMMARY OF RESULTS***
